backend/requirements.txt
cryptography==2.8
Flask==1.1.1
Flask_Cors==3.0.10
PyJWT==2.3.0
pytest==6.2.5
requests==2.22.0
Werkzeug==1.0.0